chore: 이벤트 버튼 위치 이동
